% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inat_obs_sf.R
\name{get_inat_obs_sf}
\alias{get_inat_obs_sf}
\title{Helper for rinat_get_obs to query with sf polygon object}
\usage{
get_inat_obs_sf(
  query = NULL,
  taxon_name = NULL,
  taxon_id = NULL,
  place_id = NULL,
  quality = NULL,
  geo = NULL,
  annotation = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  area = NULL,
  maxresults = 100,
  meta = FALSE
)
}
\arguments{
\item{query}{Query string for a general search.}

\item{taxon_name}{Filter by iNat taxon name. Note that this will also select observations of
descendant taxa. Note that names are not unique, so if the name matches multiple taxa, no
observations may be returned.}

\item{taxon_id}{Filter by iNat taxon ID. Note that this will also select observations of descendant taxa.}

\item{place_id}{Filter by iNat place ID.}

\item{quality}{The quality grade to be used.  Must be either "casual" or "research".  If left
blank both will be returned.}

\item{geo}{Flag for returning only results that are georeferenced, TRUE will exclude
non-georeferenced results, but they cannot be excluded.}

\item{annotation}{Filter by annotation. Vector of length 2, the first 
element being the term ID (e.g. "1" for life stage) and the second 
element being the value ID (e.g. "2" for adult). Refer to the 
\href{https://forum.inaturalist.org/t/how-to-use-inaturalists-search-urls-wiki-part-2-of-2/18792#heading--annotations}{annotation documentation} to know which values to use.}

\item{year}{Return observations only in that year (can only be one year, not a range of years).}

\item{month}{Return observations only by month, must be numeric, 1...12}

\item{day}{Return observations only on a given day of the month,  1...31}

\item{area}{Supply an sf or sp object from which the bounding box will
be derived.}

\item{maxresults}{The maximum number of results to return. Should not be
a number higher than 10000.}

\item{meta}{(logical) If TRUE, the output of this function is a list with metadata on the output
and a data.frame of the data. If FALSE (default), just the data.frame.}
}
\value{
A sf dataframe of the number of observations requested.
}
\description{
Helper written for rinat::get_inat_obs, since it usually takes a vector of coords
rather than an sf object. Often, want to query/clip by sf object for the
returned results, rather than a simple bounding box. 

for data, or just filter results by a subset of what is offered by the API.
}
\note{
Filtering doesn't always work with the query parameter for some reason (a problem on
the API end).  If you want to filter by time, it's best to use the scientific name and put it
in the 'taxa' field, and not in the query field.  Another issue is that the query parameter
will search the entire entry, so it is possible to get unintended results.  Depending on your
use case it may be advisable to use the "taxon" field instead of the query field.
}
