\name{get_obs_project}
\alias{get_obs_project}
\title{get_obs_project - function to retrive observations from a perticular
 iNaturalist project.}
\usage{
  get_obs_project(grpid)
}
\arguments{
  \item{grpid}{Name of the group or group id}
}
\description{
  get_obs_project - function to retrive observations from a
  perticular iNaturalist project.
}
\examples{
\dontrun{
 get_obs_project(354)
 get_obs_project("reptileindia")
}
}

